backend/requirements.txt
fastapi
uvicorn
aiohttp 
python-dotenv
python-multipart
sqlalchemy
asyncpg
passlib==1.7.4
bcrypt==4.0.1